(BSR)[PRO] refactor: extract RouteWrapper
